server/requirements.txt
torch
torchvision
pillow
opencv-python
tqdm
matplotlib
torchaudio
#tensorflow
pyinstaller
flask